requirements.txt
flask>=3.0
flask-login>=0.6
authlib>=1.3
mysql-connector-python>=8.4
python-dotenv>=1.0
pandas>=2.2
openpyxl>=3.1
gunicorn>=22.0
pytest>=8.0
pytest-cov>=5.0